# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente"
          },
          "400": {
            "description": "Error en validación o usuario ya existe"
          }
        },
        "summary": "Registrar nuevo usuario",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, retorna tokens y usuario",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "data": {
                      "accessToken": "eyJhbGc...",
                      "refreshToken": "v1.abc...",
                      "expiresAt": 1900000000,
                      "tokenType": "bearer",
                      "user": {
                        "id": "uuid-123",
                        "email": "user@sacdia.app",
                        "name": "Juan",
                        "paternal_last_name": "Perez",
                        "maternal_last_name": "Lopez",
                        "avatar": null,
                        "roles": [
                          "user"
                        ]
                      },
                      "needsPostRegistration": false,
                      "postRegistrationStatus": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens renovados exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "data": {
                      "accessToken": "eyJhbGc...",
                      "refreshToken": "v1.abc...",
                      "expiresAt": 1900000000,
                      "tokenType": "bearer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payload legacy no soportado (`refresh_token` retirado)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "refresh_token was removed. Use refreshToken in request body.",
                    "code": "LEGACY_SNAKE_CASE_REMOVED",
                    "removedAt": "2026-03-01",
                    "use": "refreshToken"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido o expirado"
          }
        },
        "summary": "Refrescar sesión con refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "v1.abc...",
                    "description": "Refresh token opcional para logout fail-safe cuando el access token ya expiró."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión cerrada (best effort, no bloquea UX)"
          }
        },
        "summary": "Cerrar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset-request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correo de recuperación enviado"
          }
        },
        "summary": "Solicitar recuperación de contraseña",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Información del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "string",
                          "example": "user-uuid"
                        },
                        "email": {
                          "type": "string",
                          "example": "usuario@sacdia.app"
                        },
                        "name": {
                          "type": "string",
                          "example": "Juan"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "user"
                          ]
                        },
                        "permissions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "read"
                          ]
                        },
                        "post_register_complete": {
                          "type": "boolean",
                          "example": true,
                          "description": "Indica si el usuario completó el post-registro."
                        }
                      },
                      "required": [
                        "user_id",
                        "email",
                        "roles",
                        "permissions",
                        "post_register_complete"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener perfil del usuario autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me/context": {
      "patch": {
        "operationId": "AuthController_setActiveContext_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveClubContextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contexto activo actualizado"
          },
          "400": {
            "description": "La asignación no pertenece o no está activa para el usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar contexto activo de club/instancia del usuario",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/profile/completion-status": {
      "get": {
        "operationId": "AuthController_getCompletionStatus_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estado del post-registro del usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener estado del post-registro",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Juan",
            "description": "Nombre del usuario"
          },
          "paternal_last_name": {
            "type": "string",
            "example": "García",
            "description": "Apellido paterno"
          },
          "maternal_last_name": {
            "type": "string",
            "example": "López",
            "description": "Apellido materno"
          },
          "email": {
            "type": "string",
            "example": "juan.garcia@example.com",
            "description": "Correo electrónico"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Contraseña segura (mínimo 8 caracteres, debe incluir mayúscula, minúscula, número y carácter especial)"
          }
        },
        "required": [
          "name",
          "paternal_last_name",
          "maternal_last_name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "juan.garcia@example.com",
            "description": "Correo electrónico"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Contraseña"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshSessionDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "v1.abc...",
            "description": "Refresh token (camelCase, formato soportado)"
          },
          "refresh_token": {
            "type": "string",
            "example": "v1.abc...",
            "description": "Legacy snake_case retirado el 2026-03-01. Solo habilitable temporalmente por rollback flag.",
            "deprecated": true
          }
        }
      },
      "ResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "juan.garcia@example.com",
            "description": "Correo para recuperación"
          }
        },
        "required": [
          "email"
        ]
      },
      "SetActiveClubContextDto": {
        "type": "object",
        "properties": {
          "assignment_id": {
            "type": "string",
            "description": "ID de la asignación de rol/instancia activa del usuario",
            "format": "uuid"
          }
        },
        "required": [
          "assignment_id"
        ]
      },
...
    }
  }
}